# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/interviews": {
      "post": {
        "operationId": "InterviewsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Candidate, application, user, or template not found"
          },
          "409": {
            "description": "An interviewer is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Schedule a panel interview and assign scorecards",
        "tags": [
          "Interviews"
        ]
      },
      "get": {
        "operationId": "InterviewsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interview"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List interviews",
        "tags": [
          "Interviews"
        ]
      }
    },
    "/interviews/{id}": {
      "get": {
        "operationId": "InterviewsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get interview with panel and scorecards",
        "tags": [
          "Interviews"
        ]
      }
    },
    "/interviews/{id}/reschedule": {
      "patch": {
        "operationId": "InterviewsController_reschedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleInterviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reschedule an interview with conflict detection",
        "tags": [
          "Interviews"
        ]
      }
    },
    "/interviews/{id}/outcome": {
      "patch": {
        "operationId": "InterviewsController_outcome",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewOutcomeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark an interview completed, cancelled, or no-show",
        "tags": [
          "Interviews"
        ]
      }
    },
    "/interviews/{id}/cancel": {
      "patch": {
        "operationId": "InterviewsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an interview",
        "tags": [
          "Interviews"
        ]
      }
    },
    "/interviews/scorecard-templates": {
      "post": {
        "operationId": "InterviewsController_createTemplate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScorecardTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardTemplate"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a job or stage scorecard template",
        "tags": [
          "Interviews"
        ]
      }
    },
    "/interviews/scorecard-templates/list": {
      "get": {
        "operationId": "InterviewsController_templates",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stageId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScorecardTemplate"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active scorecard templates",
        "tags": [
          "Interviews"
        ]
      }
    },
    "/interviews/scorecards/{id}/submit": {
      "post": {
        "operationId": "InterviewsController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitScorecardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scorecard"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not the assigned reviewer"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit assigned structured feedback once",
        "tags": [
          "Interviews"
        ]
      }
    },
    "/interviews/applications/{applicationId}/decision-summary": {
      "get": {
        "operationId": "InterviewsController_summary",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "applicationId",
                    "complete",
                    "missing",
                    "scorecards"
                  ],
                  "properties": {
                    "applicationId": {
                      "type": "integer"
                    },
                    "complete": {
                      "type": "integer"
                    },
                    "missing": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "scorecards": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Show completed and missing scorecards for a hiring decision",
        "tags": [
          "Interviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "GEDPro API",
    "description": "Applicant tracking API for users, candidates, jobs, applications, forms, documents, and interviews.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/v1",
      "description": "Version 1 (recommended)"
    },
    {
      "url": "/",
      "description": "Legacy unversioned compatibility"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "title should not be empty"
            ]
          },
          "path": {
            "type": "string",
            "example": "/jobs"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-15T10:00:00.000Z"
          },
          "requestId": {
            "type": "string",
            "example": "req_01JABCDEF"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "Candidate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "normalizedEmail": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "normalizedPhone": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "archivedAt": {
            "format": "date-time",
            "type": "string"
          },
          "privacyConsent": {
            "type": "boolean"
          },
          "consentAt": {
            "format": "date-time",
            "type": "string"
          },
          "retentionUntil": {
            "format": "date-time",
            "type": "string"
          },
          "deletionRequestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "mergedInto": {
            "$ref": "#/components/schemas/Candidate"
          },
          "currentState": {
            "enum": [
              "new",
              "preselected",
              "interview_scheduled",
              "in_interview",
              "accepted",
              "rejected"
            ],
            "type": "string"
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateHistory"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "normalizedEmail",
          "phone",
          "tags",
          "skills",
          "privacyConsent",
          "currentState",
          "history",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
...
      "CreateInterviewDto": {
        "type": "object",
        "properties": {
          "candidateId": {
            "type": "number",
            "minimum": 1
          },
          "applicationId": {
            "type": "number",
            "minimum": 1
          },
          "date": {
            "type": "string"
          },
          "type": {
            "enum": [
              "HR",
              "TECHNICAL",
              "FINAL"
            ],
            "type": "string"
          },
          "duration": {
            "type": "number",
            "minimum": 5,
            "maximum": 1440,
            "default": 60
          },
          "round": {
            "type": "number",
            "minimum": 1,
            "default": 1
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "location": {
            "type": "string",
            "maxLength": 500
          },
          "notes": {
            "type": "string",
            "maxLength": 5000
          },
          "interviewerIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "attendeeIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "scorecardTemplateId": {
            "type": "number",
            "minimum": 1
          },
          "feedbackDeadline": {
            "type": "string"
          },
          "hideFeedbackUntilComplete": {
            "type": "boolean"
          },
          "calendarProvider": {
            "type": "string"
          },
          "overrideConflicts": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "candidateId",
          "applicationId",
          "date"
        ]
      },
...
      "Interview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "status": {
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED",
              "RESCHEDULED",
              "NO_SHOW"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "HR",
              "TECHNICAL",
              "FINAL"
            ],
            "type": "string"
          },
          "round": {
            "type": "number"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "feedbackDeadline": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "hideFeedbackUntilComplete": {
            "type": "boolean"
          },
          "calendarProvider": {
            "type": "string",
            "nullable": true
          },
          "calendarEventId": {
            "type": "string",
            "nullable": true
          },
          "calendarSyncStatus": {
            "type": "string"
          },
          "calendarSyncError": {
            "type": "string",
            "nullable": true
          },
          "candidate": {
            "$ref": "#/components/schemas/Candidate"
          },
          "application": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Application"
              }
            ]
          },
          "interviewer": {
            "$ref": "#/components/schemas/User"
          },
          "attendees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "scorecards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Scorecard"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "date",
          "duration",
          "status",
          "type",
          "round",
          "title",
          "location",
          "notes",
          "feedbackDeadline",
          "hideFeedbackUntilComplete",
          "calendarProvider",
          "calendarEventId",
          "calendarSyncStatus",
          "calendarSyncError",
          "candidate",
          "application",
          "interviewer",
          "attendees",
          "scorecards",
          "createdAt",
          "updatedAt"
        ]
      },
      "ScorecardTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "criteria": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "job": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Job"
              }
            ]
          },
          "stage": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/PipelineStage"
              }
            ]
          },
          "archived": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "criteria",
          "job",
          "stage",
          "archived",
          "createdAt",
          "updatedAt"
        ]
      },
      "Scorecard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "interview": {
            "$ref": "#/components/schemas/Interview"
          },
          "reviewer": {
            "$ref": "#/components/schemas/User"
          },
          "template": {
            "$ref": "#/components/schemas/ScorecardTemplate"
          },
          "ratings": {
            "type": "object",
            "nullable": true
          },
          "recommendation": {
            "nullable": true,
            "enum": [
              "strong_no",
              "no",
              "yes",
              "strong_yes"
            ],
            "type": "string"
          },
          "privateNotes": {
            "type": "string",
            "nullable": true
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "interview",
          "reviewer",
          "template",
          "ratings",
          "recommendation",
          "privateNotes",
          "submittedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "RescheduleInterviewDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "duration": {
            "type": "number",
            "minimum": 5,
            "maximum": 1440
          },
          "overrideConflicts": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "date"
        ]
      },
      "InterviewOutcomeDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW"
            ],
            "type": "string"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": [
          "status"
        ]
      },
      "ScorecardCriterionDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 80
          },
          "label": {
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "minRating": {
            "type": "number"
          },
          "maxRating": {
            "type": "number"
          },
          "required": {
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "label",
          "minRating",
          "maxRating"
        ]
      },
      "CreateScorecardTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "criteria": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScorecardCriterionDto"
            }
          },
          "jobId": {
            "type": "number",
            "minimum": 1
          },
          "stageId": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "name",
          "criteria"
        ]
      },
      "SubmitScorecardDto": {
        "type": "object",
        "properties": {
          "ratings": {
            "type": "object"
          },
          "recommendation": {
            "enum": [
              "strong_no",
              "no",
              "yes",
              "strong_yes"
            ],
            "type": "string"
          },
          "privateNotes": {
            "type": "string",
            "maxLength": 10000
          }
        },
        "required": [
          "ratings",
          "recommendation"
        ]
      },
...
    }
  }
}